Preliminary support for shared upload directory. This is primarily intended for the Wikimedia Commons. If $wgUseSharedUploads is set, MediaWiki will try to locate an image file in the shared upload directory if it is not found locally. Thumbnails for shared images will also be generated in the shared directory.
Note: I changed
                $text = $this->replaceInternalLinks ( $text );
                $text = $this->replaceExternalLinks( $text );
to
                $text = $this->replaceExternalLinks( $text );
                $text = $this->replaceInternalLinks ( $text );

in Parser.php. Otherwise [[Media:Bla.jpg]] to an absolute URL would be
incorrectly parsed as external links. The header of
replaceExternalLinks says:

        Note: we have to do external links before the internal ones

So the old order was not consistent with the documentation. I did some
tests and nothing seems to be broken, but I have some problems executing
parserTests.php so I couldn't test it systematically.


git-svn-id: http://svn.wikimedia.org/svnroot/mediawiki/trunk@5989 dd0e9695-b195-4be7-bd10-2dea1a65a6b6