* (bug 4663) Edit toolbar enabled in compatible versions of Safari * (bug 5572) Edit toolbar enabled in compatible versions of Konqueror (3.5+) * (bug 5235) Edit toolbar tooltips no longer show JavaScript junk in Opera * Edit toolbar now works in pure XHTML mode (application/xhtml+xml)
The edit toolbar buttons are now created with DOM functions rather than
document.writeln(). This makes it compatible with viewing in XML mode in
strict browsers like Firefox, and additionally allows testing if the
textarea supports the 'selectionStart' property so recent versions of
Safari and Konqueror can get the toolbar automatically enabled.

Information for the toolbar items is stored in globals mwEditButtons and
mwCustomEditButtons; this latter allows MediaWiki:Common.js or similar
to quickly and easily add custom items such as the redirect button currently
on en.wikipedia.org.

Switching from a javascript: link to a straight image with an onclick
handler removes the unsightly javascript url gunk from Opera's tooltip.



git-svn-id: http://svn.wikimedia.org/svnroot/mediawiki/trunk@13689 dd0e9695-b195-4be7-bd10-2dea1a65a6b6