Spent a few spare minutes poking at the mobile gateway....
+ 2008-05-24: Page loading switched from Special:Export to API
+ 2008-05-24: Search switched from broken screen-scrape to API
+ 2008-05-24: Titles with "&" now display correctly.

Page fetching was using Special:Export but had some problems:
* Titles containing '&' weren't unescaped, so you'd see 'AT&amp;T' in the title bar
* Redirects weren't parsed correctly, like this one: http://en.wikipedia.org/w/index.php?title=Iphone&action=edit&oldid=161238452

I fixed both of those by switching to the API and letting it handle the redirects server-side. This is also slightly faster than Special:Export. :)


I'd noticed some time ago that search was really bad, but assumed it just wasn't implemented fully, doing 'go' matches only.
In fact it was screen-scraping, and changes to the HTML output of Special:Search broke it months or years ago. Nice!
Now using the API for searches, which actually returns results.

Note that the current API just throws an error unless you specify text matches only; it doesn't quite understand that our search engine has only one type of search, not separate title & text matches, and that that's ok to do. That should probably get fixed.


The README has references to some template parsing mode that uses Special:ExpandTemplates, and there are localized name references in some of the localized keywords files for it. However I see no sign of it actually existing. If this gets re-implemented, it should be done via API as well.


There remain some things such as fetching of images which should probably also be moved to the API so we're not manually constructing URLs which may break.



git-svn-id: http://svn.wikimedia.org/svnroot/mediawiki/trunk@35299 dd0e9695-b195-4be7-bd10-2dea1a65a6b6