+ 2008-05-26: Image thumb lookups now down via API for future-proofing,   correction of name normalization issues, and safety for the image.php   interface.
Old code was manually constructing a URL based on filename hashing, then checking
both the source wiki and Commons to try to find a file size, then determining whether
to use the source or a thumbnail image.

Then to top it off, the complete URL was passed to image.php, which would load
*any arbitrary URL from the internet* and attempt to convert its image type.

New code passes just the language, name, and requested width to image.php, which
does an API lookup to the target wiki to get the appropriate thumbnail URL. This
URL is then loaded and converted.

This aids in future-proofing by avoiding hard-codeded assumptions about the image
URL structure, number and location of remote repositories, file types, method of
thumbnailing, and normalization of filenames.

(Previously, images specified with a lowercase name failed to load correctly... but
on some wikis with lowercase enabled that would be legitimate as well. Leaving that
to the upstream wiki avoids having to worry about the config on the transcoder.)


Also added a fallback on image.php to do an HTTP redirect to the original thumb
if GD isn't present on the host; this is to aid debugging on my local system, since
Apple for some reaon didn't compile GD into their PHP installation.




git-svn-id: http://svn.wikimedia.org/svnroot/mediawiki/trunk@35370 dd0e9695-b195-4be7-bd10-2dea1a65a6b6