* (bug 13376) Use $wgPasswordSender, not $wgEmergencyContact, as return   address for page update notification mails.
Note that $wgEmergencyContact is now not currently in active use, but could
in theory be used for some kinds of failure notifications or to provide to
users to contact in case of failure, which I think was its original intention.
Certainly it's silly to throw an "emergency" contact on auto-sent mails!

There's also a $wgNoReplyAddress, which is currently used to add a Reply-To
header on notification mails. WTF? :)

We should probably refactor all this address crap...


git-svn-id: http://svn.wikimedia.org/svnroot/mediawiki/trunk@38374 dd0e9695-b195-4be7-bd10-2dea1a65a6b6