Revert r38591 -- "Make good-faith effort to run BrokenLink hook in Linker::link().  No parser-test regressions, and it shouldn't change behavior unless Xml::expandAttributes() and Sanitizer::decodeTagAttributes() aren't inverses up to normalization."
IMHO this is an excellent opportunity to kill a horrible interface and replace it with a sane one. Note the only use of the BrokenLink hook currently in our SVN is in SemanticForms:

/**
 * Sets the URL for form-based adding of a nonexistent (broken-linked, AKA
 * red-linked) page
 */
function sffSetBrokenLink(&$linker, $title, $query, &$u, &$style, &$prefix, &$text, &$inside, &$trail) {
	$link = sffAddDataLink($title);
	if ($link != '')
		$u = $link;
	return true;
}

In the unlikely event that anybody else is really needing the exact hook details, I'm sure there's a much nicer, more future-friendly way to do it. Make a new hook and let these hypothetical extensions fix themselves up. :)


git-svn-id: http://svn.wikimedia.org/svnroot/mediawiki/trunk@38611 dd0e9695-b195-4be7-bd10-2dea1a65a6b6