Reverting r43581 "Added a new user right called "alwaysuseskin" that allows the user to use and set their configured skin even if the wiki administrator has chosen to disable user skins for the rest of the wiki. Req. by Alexfusco5."
Seems a little weird to me; I'm not quite sure the use case. :)

Comments on the revision note that it needs documentation as well, and the name is a bit weird. If it is needed/desireable it might be better to just replace $wgAllowUserSkin with a right 'useskin' or something.


git-svn-id: http://svn.wikimedia.org/svnroot/mediawiki/trunk@43630 dd0e9695-b195-4be7-bd10-2dea1a65a6b6