Change "commentator" to "commenter"
A commentator is someone who professionally comments on things,
generally public affairs or news items.  That's certainly not correct.
Between "commentor" (the previous spelling) and "commenter", the latter
seems to be preferable: the AHD has neither, but Webster's has
"commenter" and not "commentor", and "commenter" is the clear winner in
a Google fight.


git-svn-id: http://svn.wikimedia.org/svnroot/mediawiki/trunk@44971 dd0e9695-b195-4be7-bd10-2dea1a65a6b6