Blug... commit current state; trying to find consistency in what's going on ;)
Core language files have a single $messages array...
Extension files have a $messages meta-array which contains messages for all languages, defined multiple times.
The '$messages' extraction that's used on core message files is not being applied to extensions I think, otherwise we'd only get the first one.
However, we've still got to find a clean way to deal with the 'require's and 'if's and such that the code barfs on and get the code cleaned up. Blecch


git-svn-id: http://svn.wikimedia.org/svnroot/mediawiki/trunk@55024 dd0e9695-b195-4be7-bd10-2dea1a65a6b6