Increase rows on reply form from 5 -> 10
Maybe it's just because I'm long-winded, but I always find the height to
be way too small.  Thankfully I'm using WebKit, so I can just resize it
myself, but it gets annoying.  10 seems better, maybe more would be
better still.

This isn't actually tested, since I don't have a CodeReview install
handy, but it falls in the category of "too trivial to be buggy".
For real, you know, not like all the other times when people say that
but it's actually buggy anyway.  Honest.

git-svn-id: http://svn.wikimedia.org/svnroot/mediawiki/trunk@55516 dd0e9695-b195-4be7-bd10-2dea1a65a6b6