Finish migration to TranslationHelpers.php
This remove lots of code duplication between TranslateEditAddons
and TranslationHelpers by moving missing boxes to TranslationHelpers.
This also removes any differences in boxes between the old and new
editing interface. Most notably the "in other languages" box takes
less space and is prettier.

In addition Translation memory suggestions now take advantage of the
context parameter if it exists. Suggestions now link back to the
message which was used to create the suggestion. Context parameter
is populated by the new tm-export script.

Fixes also a bug in tm suggestions querying invalid url from the
tm server if it was unable to determine language code, for example
when visiting the message page itself instead of subpage.

There is a bug in the translation page box with new js edit
interface. CSS styles for the diffs are not loaded. Possible fix
would by to load the styles always, since there isn't any mechanism
to load styles on demand afterwards.


git-svn-id: http://svn.wikimedia.org/svnroot/mediawiki/trunk@61168 dd0e9695-b195-4be7-bd10-2dea1a65a6b6