Don't break old wikis using Renameuser
This needs to be done for all the extensions that had their entry points
renamed, where possible.  (There was one I spotted that only differed in
case, and we can't have two filenames differ only in case because of all
the poor Windows users out there.)  The status quo will reliably cause
*lots* of wikis to break on update to 1.17, any that use any affected
extension.  On top of that, it causes your wiki to break whenever you
switch back and forth between trunk and a branch with the same
LocalSettings.php.

git-svn-id: http://svn.wikimedia.org/svnroot/mediawiki/trunk@66278 dd0e9695-b195-4be7-bd10-2dea1a65a6b6