Change the width of the div surrouding mediaplayers into min-width
The native a/v players of FireFox and Chrome, have a minimum width that is
rather large (220px). In many cases Wikipedia thumbs are requested at a size 
that is smaller than that. With a set width, the elements than overflow out
of fixed width divs. Using min-width instead allows the surrounding div
to resize to fit its contents.

Fixes bug 22538



git-svn-id: http://svn.wikimedia.org/svnroot/mediawiki/trunk@66505 dd0e9695-b195-4be7-bd10-2dea1a65a6b6