Delete trunk/phpwiki/
r1289 (April 2004) says this is "left here as an historical record of
the evolution of the code", but that's why we have version control to
begin with.  This sometimes confuses greps on trunk, and it serves no
purpose to keep it around.  Anyone who wants it can look in the history.

There are probably some other things in trunk/ that we can remove . . .

git-svn-id: http://svn.wikimedia.org/svnroot/mediawiki/trunk@71803 dd0e9695-b195-4be7-bd10-2dea1a65a6b6