Follow up r79417
Fix URL for email notifications.
I just added a new contextFullURL method. This part of the code is totally undocumented
and I have 0 idea what *Url function is supposed to be doing what at this stage. It 
needs cleanup, but I'm not familiar enough with the code to be able to clean this up any further.


git-svn-id: http://svn.wikimedia.org/svnroot/mediawiki/trunk@79608 dd0e9695-b195-4be7-bd10-2dea1a65a6b6