(follow-up r83433) Fix (or workaround) bug discovered with mode=none having <pre>'s inserted (root cause is bug 1319)
Fix up the tests I added. Some of them were testing stuff thats not going to work
(such as the addfirstcategorydate, since the date is going to change from day to day. d'oh).
Remove the one whitespace before the li on all these tests (for the pre issue). And add one more test.


git-svn-id: http://svn.wikimedia.org/svnroot/mediawiki/trunk@83621 dd0e9695-b195-4be7-bd10-2dea1a65a6b6