Make it so that when a special page is trancluded, the output won't vary by url parameter.
This stops severe ugliness http://test.wikipedia.org/wiki/User:Bawolff/special?feed=atom
where the rss feed and the html of the page is concated together.

This could potentially break stuff if someone was using a transcludable special page
with an html form, or if someone is actually passing parameters to the transcludable
special page via the url. I'm not aware of anyone who does that, and both those things
seem rather evil.


git-svn-id: http://svn.wikimedia.org/svnroot/mediawiki/trunk@87705 dd0e9695-b195-4be7-bd10-2dea1a65a6b6