Initial start on batch tests for ParserPlayground. Wrapper to do round-trip testing (parse, then return to source), running the tests on input from a MediaWiki XML export dump.
Won't be very successful at the moment as newlines between paragraphs are not properly preserved. :)

Added commonJS-style exports to a couple of the PP JS modules so they're loadable within node.js.

Tested with node.js 0.4.9 & npm package manager -- README indicates which modules need to be installed with the npm package manager.



git-svn-id: http://svn.wikimedia.org/svnroot/mediawiki/trunk@92128 dd0e9695-b195-4be7-bd10-2dea1a65a6b6