back off r94558:  - reverts ./includes/ProxyTools.php  - marks tests broken
In our test suite, the first call to wfGetIP() set the static variable.
Hence the remaining of the code is only tested on the first call to it.
Resetting the static variable enlight a bug somewhere in our code where
we are calling wfGetIP() but can not reliably get an IP, somehow REMOTE_ADDR
does not exist and $wgCommandLineMode is disabled.
Will have to track this bug further when I got time.



git-svn-id: http://svn.wikimedia.org/svnroot/mediawiki/trunk@94577 dd0e9695-b195-4be7-bd10-2dea1a65a6b6