add ability to attach a comment to a diff line
cc_patch_line === NULL means the comment is below the diff just like it
was before.  Set to a number, it refers to the line number in the diff
file from 1 to n.

Still need an UI which is a work in progress.

NOTE: adding a simple field required a lot of files to be changed :(

git-svn-id: http://svn.wikimedia.org/svnroot/mediawiki/trunk@95429 dd0e9695-b195-4be7-bd10-2dea1a65a6b6