(bug 30617) inline comments forms cannot be closed
On clicking a line, a javascript insert a <tr> element containg the
form. That <tr> did not get an unique id which caused troubles when
opening multiples inline comments.

This patch forge an id based on the diff line it applies to which should
be unique since we only allow one comment form per line.

git-svn-id: http://svn.wikimedia.org/svnroot/mediawiki/trunk@95680 dd0e9695-b195-4be7-bd10-2dea1a65a6b6