* (bug 30900) Remove confusing default value added for the default-don't-say-anything case of searchmenu-new-nocreate message
Reverts r80746's addition of a default value, removal of the message from the 'ignore' list, and the subsequent translations of this message that's supposed to be empty.
It doesn't make sense to have a default value here; as added in r65902 it was meant to default to empty. It's not very well thought out to begin with, so any message that goes it ends up being confusing. For queries that aren't meant to be page titles it's meaningless to begin with, and for users who aren't allowed to create articles at all due to read-only or edit/create restrictions it's even more confusing.




git-svn-id: http://svn.wikimedia.org/svnroot/mediawiki/trunk@97098 dd0e9695-b195-4be7-bd10-2dea1a65a6b6