* (bug 31089) Fix regression in Special:Block reasons list (MediaWiki:Ipbreason-dropdown parsing)
Followup to r83298: makes HTMLSelectAndOtherField backwards-compatible with Xml::listDropDown as it was apparently intended to be.

Also fixed double-escaping in the options for HTMLSelectAndOtherField.

Followup to r83795: fixes regression in Special:Block by fixing HTMLSelectAndOtherField which got used in place of Xml::listDropDown, but wasn't compatible.


Note that this removes the clever thing that lets you have separate values and texts in your dropdown; since Xml::listDropDown had no such feature and actual usage includes texts containing links and other stuch where the pipe character is a legit part of the content.




git-svn-id: http://svn.wikimedia.org/svnroot/mediawiki/trunk@97860 dd0e9695-b195-4be7-bd10-2dea1a65a6b6