Commit some more of this half-rewritten code before I lose it. :)
Expansion bits half-ported from the PHP, needs to be finished & the formats normalized a bit.
Attempting to model the template expansion in a node tree on PPFrame::expand from preprocessor, but outputting another node tree (this is broken right now) rather than string output.
It needs to be able to exit and re-enter later when fetching templates asynchronously, though it doesn't use a very efficient plan yet (just waits).



git-svn-id: http://svn.wikimedia.org/svnroot/mediawiki/trunk@97979 dd0e9695-b195-4be7-bd10-2dea1a65a6b6