Suppress annoying PHP 5.3 warning due to use of PHP4-style references.
There are a *lot* more reference misuses in this extension that should be cleaned up, but I can't actually get the ext to fully work. On my trunk/1.19 test instance, with a maps API key manually set up, I see only a vertical space where the map should be. Some images and scripts get loaded, but not inserted... possibly something's incompatible with how we've changed some of the JS setup.

I'd also recommend that someone interested update this to Google Maps API v3 -- this doesn't require an API key for low-level usage, so should be easier to set up.



git-svn-id: http://svn.wikimedia.org/svnroot/mediawiki/trunk@99567 dd0e9695-b195-4be7-bd10-2dea1a65a6b6