The following patches restore how texvc added braces for all commands but \operatorname to fix Bug 31442.  More specifically, TEX_FUN1 now adds braces as it did in previous versions.  The result is texvc will once again correctly sanitize multiple math accents even when no braces are given (e.g. \dot \vec B) and it will allow math accents to be used with symbols whose font as been changed (e.g. \tilde \mathcal{M}).  Parser tests are created to ensure these continue to work.
The handling of \operatorname is now  handled specially by the
function TEX_FUN1nb (nb stands for no braces and the name is taken to
be consistent with the existing function TEX_FUN2nb).  The addition of
braces causes this command to LaTeX incorrectly.  The parser test for
\operatorname has been updated to reflect the fact the hash of the
images involved has changed.   
  


git-svn-id: http://svn.wikimedia.org/svnroot/mediawiki/trunk@99741 dd0e9695-b195-4be7-bd10-2dea1a65a6b6