(bug 31656) AbuseFilter skips non null edits
This bug is triggering when a user submit an old revision again. AbuseFilter
considered it to be a null edit although the old revision is certainly
a different text than the current one in the database.

This patch make sure we skip filtering only if we will be comparing with the
lastest content.

Follow up r52740


git-svn-id: http://svn.wikimedia.org/svnroot/mediawiki/trunk@100687 dd0e9695-b195-4be7-bd10-2dea1a65a6b6