Data handling cleanup in DonationInterface, part 1 of [some]. in GatewayAdapter: Renamed getData_Raw to getData_Unstaged_Escaped Renamed the $raw_data member to $unstaged_data
in DonationData: 
Created public functions getDataEscaped, getDataUnescaped, and getVal_Escaped
Changed normalizeAndSanitize() so it will only normalize (with a name change to just "normalize") and moved the call to htmlspecialchars to a new public getDataEscaped function. (More visibility there)
Because of the changes in the escaping, getVal now returns the unescaped value, and is protected for internal use only. 
Protected many functions intended only to be used by their containing object. 

Renamed all references to renamed functions and data members accordingly. 
Cleaned up some logging we don't have to do anymore, and added some we do.

git-svn-id: http://svn.wikimedia.org/svnroot/mediawiki/trunk@105938 dd0e9695-b195-4be7-bd10-2dea1a65a6b6